requirements.txt
wheel==0.36.2
Flask==1.1.2
environs==9.3.1
python-telegram-bot==13.2
flask==sslify-0.1.5
redis==3.5.3
fake-useragent==0.1.11
beautifulsoup4==4.9.3
lxml==4.6.2
more-itertools==8.7.0